Initial commit of the translation plugin
This comes with a file that builds an xml with contents of all posts
and with a plugin to wordpress that uses xml2po to convert this xml
to a po file. This also contains hooks to the_{title,content} that
uses gettext() to get translated versions of the site content.